# api/openapi.yaml
openapi: 3.1.0
info:
  title: Product Video Studio API
  version: 0.1.0
servers:
  - url: /api
paths:
  /auth/login:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginInput' }
      responses:
        '200': { description: Authenticated }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /auth/me:
    get:
      responses:
        '200': { description: Current user, content: { application/json: { schema: { $ref: '#/components/schemas/User' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes: { extensionToken: { type: http, scheme: bearer } }
...
  responses: { Unauthorized: { description: Not authenticated } }
  schemas:
    LoginInput: { type: object, required: [email, password], properties: { email: { type: string, format: email }, password: { type: string, format: password } } }
    User: { type: object, required: [id, email, display_name, role], properties: { id: { type: string, format: uuid }, email: { type: string }, display_name: { type: string }, role: { enum: [admin, member] } } }
...